fix unit test

+++ src/Compilers/CSharp/Test/IOperation/IOperation/IOperationTests_ISwitchOperation.cs
@@ -1469,9 +1469,9 @@ public sealed class MyClass
 }
 ";
             var expectedDiagnostics = new[] {
-                // (8,18): error CS9098: Cannot implicitly convert type 'int' to type 'MyClass' using user-defined conversion in pattern expression
+                // (8,18): error CS9098: Cannot implicitly convert type 'long' to type 'MyClass' using user-defined conversion in pattern expression
                 //             case 1:
-                Diagnostic(ErrorCode.ERR_NonConstantConversionInConstantPattern, "1").WithArguments("int", "MyClass").WithLocation(8, 18)
+                Diagnostic(ErrorCode.ERR_NonConstantConversionInConstantPattern, "1").WithArguments("long", "MyClass").WithLocation(8, 18)
             };
 
             string expectedFlowGraph = @"
